Rename entry point
Rename the entry point to be `main.js`.
